requirements.txt
semantic-version
requests
python-dotenv
pydantic
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
